packages/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Example Django backend for yetAnotherBoilerplate"
authors = ["Your Team <dev@example.com>"]
readme = "README.md"
packages = [
  { include = "boilerplate" },
  { include = "apps", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14"
django-environ = "^0.11.2"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.9"
# Optional ASGI server if you prefer running ASGI locally
uvicorn = { version = "^0.30.0", optional = true }

gunicorn = "^21.2.0"
django-anymail = "^11.0"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
pytest = "^8.3.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["boilerplate", "apps"]
force_single_line = false
combine_as_imports = true
include_trailing_comma = true
known_first_party = ["boilerplate", "apps"]
skip_gitignore = true

[tool.ruff]
# Match project Python version and Black's default line length
target-version = "py311"
line-length = 88
extend-exclude = [
  "**/migrations/**",
]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # import sorting
  "UP", # pyupgrade
  "B",  # bugbear
]
ignore = [
  "E501",  # ignore line-length to match Black defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
